@@ -6,11 +6,11 @@
     'configure_icon' => 'fa fa-sliders',
     'default_domain' => 'Default domain:',
     'default_domain_help' => 'The domain name is configured in the Site System Configuration, Site tab. This domain will be used if the incoming request does not match any of the following.',
+    'description' => 'Setting up additional site domains.',
     'domain' => 'Domain host:',
     'domain_help' => 'Enter a domain host without protocol.',
     'domain_on' => 'Enable domain:',
     'domain_on_help' => 'If the domain is disabled, access to the pages of this domain will be prohibited.',
-    'description' => 'Setting up additional site domains.',
     'empty_field' => 'the field must not be empty.',
     'hide_from_tree' => 'Hide in the resource tree:',
     'hide_from_tree_help' => 'If this box is checked, domain resources will not be displayed in the resource tree.',
