@@ -6,11 +6,11 @@
     'configure_icon' => 'fa fa-sliders',
     'default_domain' => 'Домен по умолчанию:',
     'default_domain_help' => 'Название домена настраивается в Системной конфигурации сайта, таб Сайт. Этот домен будет задействован, если входной запрос не отвечает ни одному из тех, что приведены ниже.',
+    'description' => 'Настройка дополнительных доменов сайта.',
     'domain' => 'Доменное имя:',
     'domain_help' => 'Укажите доменное имя без протокола.',
     'domain_on' => 'Включить домен:',
     'domain_on_help' => 'Если домен выключен, то доступ к страницам этого домена будет запрещен.',
-    'description' => 'Настройка дополнительных доменов сайта.',
     'empty_field' => 'поле не должно быть пустым.',
     'hide_from_tree' => 'Спрятать в дереве ресурсов:',
     'hide_from_tree_help' => 'Если это поле будет отмечено, то ресурсы домена не будут отражены в дереве ресурсов.',
