@@ -6,11 +6,11 @@
     'configure_icon' => 'fa fa-sliders',
     'default_domain' => 'Домен типово:',
     'default_domain_help' => 'Назва домену налаштовується в Системній кофігурації сайту, таб Сайт. Цей домен буде задіяно, якщо вхідний запит не відповідає жодному з тих, що наведено нижче.',
+    'description' => 'Налаштування додаткових доменів сайту.',
     'domain' => 'Доменне ім`я:',
     'domain_help' => 'Вкажіть доменне ім`я без протоколу.',
     'domain_on' => 'Ввімкнути домен:',
     'domain_on_help' => 'Якщо домен вимкнено, то доступ до сторінок цього домену буде заборонено.',
-    'description' => 'Налаштування додаткових доменів сайту.',
     'empty_field' => 'поле не має бути порожнім.',
     'hide_from_tree' => 'Приховати в дереві ресурсів:',
     'hide_from_tree_help' => 'Якщо це поле буде відмічено, то ресурси домену не буде відображено в дереві ресурсів.',
