@@ -1,5 +1,5 @@
 <style>
-    #copyright{position:fixed;bottom:0;right:0;}
-    #copyright img{width:35px;}
-    .badge.bg-seigerit{background-color:#0057B8 !important;color:#FFD700;font-size:85%;}
+    #copyright{position:fixed;bottom:0;right:0;padding:10px;}
+    #copyright img{width:35px;height:auto;}
+    .badge.bg-seigerit{background-color:#0057B8 !important;color:#FFD700;font-size:85%;padding:2px 5px;border-radius:5px;}
 </style>
